Ditch item positioning algorithm and implement manual positioning
Since the pipeline is static - as in, the user is not able to add new elements to the pipeline - there is no reason to develop a convoluted algorithm for computing the placement of the items.
A nicer, and arguably simpler method, is to do a combination of manual placement of the items, and do item placement in relation to each other.