Bugfix: LH/SH would be ignored cause of incorrect register signal interpretation #21
Previously, we interpreted the register value as a boolean - boolean register (signal) casting is interpreted as masking the signal value by 0b1. Since the r_MemRead_EXMEM value for loading a halfword is 0b10, this would be ignored.

This commit uses the correct cast for identifying whether to read memory