Stop pipeline "run" execution when pipeline signals finished #31
Run execution would not check for ecall::exit within the pipeline, and thus
would only stop when there were no more instructions to fetch and execute.
While execution is correct, this results in a bug where a discrepancy
between the cycles and instructions executed between stepping through the
program and running through the program, were different.