Add RipesLib as a PRE_TARGETDEPS library in .pro files
Previously, changes in RipesLib would not cause a relink of the app and test
projects. This is fixed by adding the library to the PRE_TARGETDEPS variable.

Furthermore, the library inclusion code is cleaned up in accordance to the qt
way of static library inclusion: https://doc.qt.io/qtcreator/creator-project-qmake-libraries.html