Initial commit for handling program finishing sequence
Some more work needs to be done wrt. "nicely" clearing the pipeline when finishing
the execution of a program. However, this commit lays the groundwork for a
handshake protocol between the outside environment and the RipesProcessor,
wherein the outside environment will request the processor to finish, and the
processor emitting a signal once it determines that it has finished.