Add unit test source for ELF testing
The file is not added to the test suite as of yet.
Compiling ELF files requires the baremetal RISC-V toolchain to be
available. This is not yet available on the CI test machines. As such,
a unit test will be added once CI has access to a prebuilt instance of
the toolchain.