Add FinalizeReason to indicate why a processor must finalize
There is a difference between finalizing if a processor fetches
instructions outside of the executable region and if the processor
has executed an exit system call.

For the latter, it is not possible to return to normal execution,
and hence the pipeline may only be executed additionally for the
given number of cycles it takes to clear the pipeline.