Fix error in assembling after resetting codeeditor
An issue occurred wherein, if a syntax error existed on a given line, and the
entire source code was subsequently deleted, then the error for the violating
line would not be invalidated by the syntaxhighlighter.
The syntaxhighlighter only updates error strings for blocks present in the
document, hence not blocks which are no longer present.
With this change, all errors for blocks outside of the current number of
blocks in the document, are removed.