Handle asynchronous trapping to execution environment
Previously, a crash could occur if the processor was 'Run' but
encountered an error whilst trapping. This was due to the fact that an
error handler may prompt a GUI response - which is invalid, given that
the trapping was performed from the simulator thread.

This commit addresses this, ensuring that control is properly handed
off to the GUI thread as well as synchronizing the GUI and simulator
threads.