ISA exts. of processors may now be toggled
Processors announce the set of supported ISA extensions.
These are then reflected in the processor selection dialog.
Whatever extensions are selected are then passed to the processor
during instantiation, which is then expected to only supported
the base + enabled ISAs. Furthermore, this set of enabled
extensions is then also used in instantiating the corresponding
assembler.