Add cosimulation test case
For a given test program, executes a reference model (RVSS) to generate a trace of register modifications. Then, the target processor is simulated, and its register trace is compared to the reference trace. From this, we can detect whether (and where) register state divergence occurs, which indicates an error in a processor implementation