Fully expand match trees #166
We need to match on all provided op parts, even if there is no aliasing purely based on comparisons with the opcodes of other instructions.

It might be that different instructions alias across non-opcode fields such as immediates or registers. If the ISA is valid, matching on all opcode parts should yield the correct instruction.