Fix crash when using extended layout for 5-stage processor.
Prior to this commit, switching to the extended layout for the 5-stage
processor would crash Ripes with a 'std::out_of_range' exception caused
by `map::at`.
This renders Ripes unusable until `processor_layout_id=1` was removed or
set to `0` in the configuration as Ripes would also crash on startup.