Add wasHit/wasMiss cache data, and fix moving average
Moving average was calculated incorrectly, as shown in #251. I can't recall the exact reason for the old implementation, but my hunch is that i wanted the moving average to be able to be plotted in the "hits/total access" configuration, which led to this notion of tracking last iteration differences etc. etc..
However, this methodology is flawed to begin with, since the moving average (obviously, in hindsight) need to be calculated on a a range of "was cycle # a hit or miss" values across a given window.

A moving average can now be achieved by setting "was hit / 1" as the configuration, and enabling the moving average (and disabling the "ratio" plot). I still think allowing a user to freely mix and max numerators and denominators to their liking is a good design descision, however, i could also easily imagine that some people wouldn't immediatly intuit themselves into the exact configuration needed for a moving average to show.
Secondly, i'd like to have the option of superimposing multiple plots - this would be needed to see e.g. the relation between moving average hitrate and total hitrate.

Both of these features need a bit more work to be implemented correctly, so that will have to be in a separate, followup commit.