Go back to virtual function-based ISAInfoBase implementation
I tried to get all the ISA info to be implemented in static functions so
they can be called from anywhere, but it ended up not working well. When
the ISA needs extensions to be defined at compile-time, then all the
possible combinations of extensions would also need to be defined at
compile-time. This proved to be too inconvenient, so the extensions are
defined as a runtime vector instead.