Add RISC-V ecall test for file operations
Updated `postToGUIThread()` to only run the passed function if the
Qt app is initialized. This was causing problems when using ecall for
file operations in `tst_riscv.cpp` becuse it would try to use the
uninitialized Qt app (a null pointer) to access the GUI thread during
syscall execution.

Added syscall execution to `tst_riscv.cpp`.