Start reworking UI after learning more about Qt
The sliderules tab will now use two "models" that contains encoding
and decoding data. Custom Qt widgets and the Qt signal/slot mechanism
will be used for all data/filter updates.

For example, a signal that is emitted when the encoding table's ISA
family is changed is connected only to a slot in the encoding model. The
model would then emit its own signal that is connected to any UI slot
that needs to display the changed data. This flow of `User Interaction ->
Model Update -> UI Update` will ensure there are no recursive loops in the
signal/slot connection tree.